Simplify and inline use-data logic
